refactor : (#32) modify folder structure
