+++ src/main/kotlin/entry/dsm/gitauth/equusgithubauth/domain/auth/entity/repository/RefreshTokenRepository.kt
@@ -1,0 +1,6 @@
+package entry.dsm.gitauth.equusgithubauth.domain.auth.entity.repository
+
+import entry.dsm.gitauth.equusgithubauth.domain.auth.entity.RefreshToken
+import org.springframework.data.repository.CrudRepository
+
+interface RefreshTokenRepository : CrudRepository<RefreshToken, String>
