+++ src/main/kotlin/entry/dsm/gitauth/equusgithubauth/domain/auth/service/SignUpService.kt
@@ -3,7 +3,7 @@ package entry.dsm.gitauth.equusgithubauth.domain.auth.service
 import entry.dsm.gitauth.equusgithubauth.domain.auth.presentation.dto.request.SignUpRequest
 import entry.dsm.gitauth.equusgithubauth.domain.user.entity.User
 import entry.dsm.gitauth.equusgithubauth.domain.user.entity.repository.UserRepository
-import entry.dsm.gitauth.equusgithubauth.domain.user.presentation.dto.response.TokenResponse
+import entry.dsm.gitauth.equusgithubauth.domain.auth.presentation.dto.response.TokenResponse
 import entry.dsm.gitauth.equusgithubauth.global.security.jwt.JwtTokenProvider
 import org.springframework.security.crypto.password.PasswordEncoder
 import org.springframework.stereotype.Service
