--- src/main/kotlin/entry/dsm/gitauth/equusgithubauth/domain/user/entity/repository/RefreshTokenRepository.kt
@@ -1,6 +1,0 @@
-package entry.dsm.gitauth.equusgithubauth.domain.user.entity.repository
-
-import entry.dsm.gitauth.equusgithubauth.domain.user.entity.RefreshToken
-import org.springframework.data.repository.CrudRepository
-
-interface RefreshTokenRepository : CrudRepository<RefreshToken, String>
